Refactor(notice): 게시글, 게시글 작성 컴포넌트 구조 수정
