+++ src/app/(main)/notice/create-post/_components/PostHero.tsx
