+++ src/app/(main)/notice/create-post/_components/index.ts
@@ -1,0 +1,2 @@
+export { CreateNoticePostHero } from './PostHero'
+export { CreateNoticePostForm } from './Form'
