+++ src/app/(main)/notice/posts/[postId]/_components/PostDetail.tsx
