+++ src/app/(main)/notice/posts/[postId]/_components/PostHero.tsx
