+++ src/app/(main)/notice/posts/[postId]/_components/index.ts
@@ -1,0 +1,2 @@
+export { NoticePostDetail } from './PostDetail'
+export { NoticePostHero } from './PostHero'
