+++ src/app/(main)/notice/posts/[postId]/page.tsx
@@ -1,5 +1,11 @@
-import { NoticePostHero } from './_components/NotciePostHero'
-import { NoticePostSection } from './_components/NoticePostSection'
+'use client'
+
+import { BoardNavigationButton } from '@/components/PostView/BoardNavigationButton'
+import PostContent from '@/components/PostView/PostContent'
+import { Spinner } from '@/components/Spinner'
+import { useGetPost } from '@/service/data/post'
+
+import { NoticePostDetail, NoticePostHero } from './_components'
 
 type NoticePostPageParams = {
   params: {
@@ -8,10 +14,21 @@ type NoticePostPageParams = {
 }
 
 const NoticePostPage = ({ params }: NoticePostPageParams) => {
+  const { data: post, status } = useGetPost({ postId: Number(params.postId) })
+
+  if (status === 'pending' || !post)
+    return (
+      <div className="flex justify-center pt-10">
+        <Spinner />
+      </div>
+    )
+
   return (
     <div>
       <NoticePostHero />
-      <NoticePostSection postId={Number(params.postId)} />
+      <NoticePostDetail post={post} />
+      <PostContent content={post.postContent} />
+      <BoardNavigationButton />
     </div>
   )
 }
